move derived dag compiler to its own file
