+++ src/DagGenerator.cpp
@@ -732,281 +732,6 @@ bool WriteSharedResources(const JsonArrayValue *resources, BinarySegment *main_s
     return true;
 }
 
-static void SortBufferOfFileAndHash(Buffer<FileAndHash>& buffer)
-{
-    std::sort(buffer.begin(), buffer.end(), [](const FileAndHash& a, const FileAndHash& b) { return strcmp(a.m_Filename, b.m_Filename) < 0; });
-}
-
-struct CompileDagDerivedWorker
-{
-    BinaryWriter _writer;
-    BinaryWriter* writer;
-    HashTable<CommonStringRecord, kFlagCaseSensitive> shared_strings;
-
-    BinarySegment *main_seg;
-    BinarySegment *data_seg;
-    BinarySegment *arraydata_seg;
-    BinarySegment *arraydata2_seg;
-    BinarySegment *leafnodearray_seg;
-    BinarySegment *scannerindex_seg;
-    BinarySegment *str_seg;
-
-    DagRuntimeData dagRuntimeData;
-    const Frozen::Dag* dag;
-    MemAllocHeap*  heap;
-    MemAllocLinear* scratch;
-    int node_count;
-
-    void WriteFrozenArrayOfBackLinks()
-    {   //write backlinks array of arrays
-        Buffer<int32_t> *links = HeapAllocateArrayZeroed<Buffer<int32_t>>(heap, node_count);
-        for (int32_t i = 0; i < node_count; ++i)
-        {
-            for(int dep : dag->m_DagNodes[i].m_Dependencies)
-            BufferAppendOne(&links[dep], heap, i);
-        }
-        BinarySegmentWriteUint32(main_seg, node_count);
-        BinarySegmentWritePointer(main_seg, BinarySegmentPosition(data_seg));
-        for (int32_t i = 0; i < node_count; ++i)
-        {
-            BinarySegmentWriteInt32(data_seg, links[i].m_Size);
-            BinarySegmentWritePointer(data_seg, BinarySegmentPosition(arraydata_seg));
-            for(int32_t backLink : links[i])
-                BinarySegmentWriteInt32(arraydata_seg, backLink);
-        }
-        for (size_t i = 0; i < node_count; ++i)
-            BufferDestroy(&links[i], heap);
-        HeapFree(heap, links);
-    }
-
-
-    struct ScannerIndexWithListOfFiles
-    {
-        int32_t m_ScannerIndex;
-        Buffer<FileAndHash> m_FilesToScan;
-    };
-
-    struct PerNodeWorkerData
-    {
-        Buffer<FileAndHash> leafInputBuffer;
-        HashSet<kFlagPathStrings> leafInputsCollector;
-        Buffer<ScannerIndexWithListOfFiles> scannersWithListsOfFiles;
-        Buffer<int32_t> all_dependent_nodes;
-        const Frozen::DagNode* dagNode;
-    };
-
-    void PerNodeWorkerDataInit(PerNodeWorkerData* data)
-    {
-        BufferInit(&data->scannersWithListsOfFiles);
-        BufferInit(&data->leafInputBuffer);
-        BufferInit(&data->all_dependent_nodes);
-        HashSetInit(&data->leafInputsCollector, heap);
-    }
-
-    void PerNodeWorkerDataDestroy(PerNodeWorkerData* data)
-    {
-        BufferDestroy(&data->scannersWithListsOfFiles, heap);
-        BufferDestroy(&data->leafInputBuffer, heap);
-        BufferDestroy(&data->all_dependent_nodes, heap);
-        HashSetDestroy(&data->leafInputsCollector);
-    }
-
-    PerNodeWorkerData m_PerNodeWorkerData;
-
-    void AddToLeafInputs(const FrozenFileAndHash& file, const Frozen::DagNode& childDagNode)
-    {
-        const Frozen::DagNode* fileGeneratingNode;
-        if (FindDagNodeForFile(&dagRuntimeData, file.m_FilenameHash, file.m_Filename.Get(), &fileGeneratingNode))
-        {
-            //we skip generated files, but we will add their FilesThatMightBeIncluded
-            if (fileGeneratingNode)
-            {
-                for(auto& f: fileGeneratingNode->m_FilesThatMightBeIncluded)
-                    AddToLeafInputs(f, *fileGeneratingNode);
-            } else {
-                //this can be true if we do know a file is generated, but we do not know which node made it. Only happens with DirectoriesCreatingImplicitDependencies,
-                //which we should remove.
-            }
-
-            return;
-        }
-
-        //did we already process this?
-        if (!HashSetInsertIfNotPresent(&m_PerNodeWorkerData.leafInputsCollector, file.m_FilenameHash, file.m_Filename.Get()))
-            return;
-
-
-        FileAndHash leafInputFile;
-        leafInputFile.m_Filename = file.m_Filename.Get();
-        leafInputFile.m_FilenameHash = file.m_FilenameHash;
-
-        BufferAppendOne(&m_PerNodeWorkerData.leafInputBuffer, heap, leafInputFile);
-
-        //if there's no scanner, we're done.
-        if (childDagNode.m_ScannerIndex == -1)
-            return;
-
-        auto findOrMakeScannerWithListOfFilesFor = [&](int scannerIndex) -> ScannerIndexWithListOfFiles*
-        {
-            for (auto& s: m_PerNodeWorkerData.scannersWithListsOfFiles)
-                if (s.m_ScannerIndex == scannerIndex)
-                    return &s;
-            ScannerIndexWithListOfFiles* newEntry = BufferAlloc(&m_PerNodeWorkerData.scannersWithListsOfFiles, heap, 1);
-            newEntry->m_ScannerIndex = scannerIndex;
-            BufferInit(&newEntry->m_FilesToScan);
-            return newEntry;
-        };
-
-        ScannerIndexWithListOfFiles* scannerIndexWithListsOfFiles = findOrMakeScannerWithListOfFilesFor(childDagNode.m_ScannerIndex);
-        BufferAppendOne(&scannerIndexWithListsOfFiles->m_FilesToScan, heap, leafInputFile);
-    }
-
-    void WriteLeafInputsAndScannerIndicesToFilesToScan_ForSingleNode(int nodeIndex)
-    {
-        const Frozen::DagNode& node = dag->m_DagNodes[nodeIndex];
-        if (0 == (node.m_Flags & Frozen::DagNode::kFlagCacheableByLeafInputs))
-        {
-            BinarySegmentWriteInt32(leafnodearray_seg, 0);
-            BinarySegmentWriteNullPointer(leafnodearray_seg);
-
-            BinarySegmentWriteInt32(scannerindex_seg, 0);
-            BinarySegmentWriteNullPointer(scannerindex_seg);
-            return;
-        }
-
-        PerNodeWorkerDataInit(&m_PerNodeWorkerData);
-
-        FindDependentNodesFromRootIndex(heap, dag, nodeIndex, m_PerNodeWorkerData.all_dependent_nodes);
-
-        for(int32_t childNodeIndex : m_PerNodeWorkerData.all_dependent_nodes)
-        {
-            const Frozen::DagNode& childDagNode = dag->m_DagNodes[childNodeIndex];
-
-            for (auto& inputFile: childDagNode.m_InputFiles)
-                AddToLeafInputs(inputFile, childDagNode);
-            for (auto& fileThatMightBeIncluded: childDagNode.m_FilesThatMightBeIncluded)
-                AddToLeafInputs(fileThatMightBeIncluded, childDagNode);
-        }
-
-        SortBufferOfFileAndHash(m_PerNodeWorkerData.leafInputBuffer);
-
-        BinarySegmentWriteInt32(leafnodearray_seg, m_PerNodeWorkerData.leafInputBuffer.m_Size);
-        BinarySegmentWritePointer(leafnodearray_seg, BinarySegmentPosition(arraydata_seg));
-        for(const FileAndHash& leafInput: m_PerNodeWorkerData.leafInputBuffer)
-        {
-            WriteCommonStringPtr(arraydata_seg, str_seg, leafInput.m_Filename, &shared_strings, scratch);
-            BinarySegmentWriteInt32(arraydata_seg, leafInput.m_FilenameHash);
-        }
-
-        BinarySegmentWriteInt32(scannerindex_seg, m_PerNodeWorkerData.scannersWithListsOfFiles.m_Size);
-        BinarySegmentWritePointer(scannerindex_seg, BinarySegmentPosition(arraydata_seg));
-        for(ScannerIndexWithListOfFiles& scannerIndexWithListOfFiles: m_PerNodeWorkerData.scannersWithListsOfFiles)
-        {
-            BinarySegmentWriteInt32(arraydata_seg, scannerIndexWithListOfFiles.m_ScannerIndex);
-            BinarySegmentWriteInt32(arraydata_seg, scannerIndexWithListOfFiles.m_FilesToScan.m_Size);
-            BinarySegmentWritePointer(arraydata_seg, BinarySegmentPosition(arraydata2_seg));
-
-            SortBufferOfFileAndHash(scannerIndexWithListOfFiles.m_FilesToScan);
-            for(const FileAndHash& fileForScanner: scannerIndexWithListOfFiles.m_FilesToScan)
-            {
-                WriteCommonStringPtr(arraydata2_seg, str_seg, fileForScanner.m_Filename, &shared_strings, scratch);
-                BinarySegmentWriteInt32(arraydata2_seg, fileForScanner.m_FilenameHash);
-            }
-            BufferDestroy(&scannerIndexWithListOfFiles.m_FilesToScan, heap);
-        }
-
-        PerNodeWorkerDataDestroy(&m_PerNodeWorkerData);
-    }
-
-
-    void WriteLeafInputsAndScannerIndicesToFilesToScan()
-    {
-        DagRuntimeDataInit(&dagRuntimeData, dag, heap);
-
-        //This function writes two arrays in parallel. we first write the headers for both, and make sure to
-        //make sure to use two different payload segments:
-
-        //Write LeafInputs array header
-        BinarySegmentWriteUint32(main_seg, node_count);
-        BinarySegmentWritePointer(main_seg, BinarySegmentPosition(leafnodearray_seg));
-
-        //Write m_ScannerIndices_To_Files_To_Scan header
-        BinarySegmentWriteUint32(main_seg, node_count);
-        BinarySegmentWritePointer(main_seg, BinarySegmentPosition(scannerindex_seg));
-
-        for (int32_t nodeIndex = 0; nodeIndex < node_count; ++nodeIndex)
-        {
-            WriteLeafInputsAndScannerIndicesToFilesToScan_ForSingleNode(nodeIndex);
-        }
-
-        DagRuntimeDataDestroy(&dagRuntimeData);
-    }
-
-    void WriteLeafInputHashOffline()
-    {
-        Buffer<int32_t> all_dependent_nodes;
-        BufferInitWithCapacity(&all_dependent_nodes, heap, 1024);
-        {  //Write m_LeafInputHash_OffLine
-            BinarySegmentWriteUint32(main_seg, node_count);
-            BinarySegmentWritePointer(main_seg, BinarySegmentPosition(data_seg));
-            for (int32_t i = 0; i < node_count; ++i)
-            {
-                HashDigest hashResult = {};
-
-                const Frozen::DagNode& node = dag->m_DagNodes[i];
-                if (0 != (node.m_Flags & Frozen::DagNode::kFlagCacheableByLeafInputs))
-                {
-                    hashResult = CalculateLeafInputHashOffline(dag, i, heap, &all_dependent_nodes);
-                }
-                BinarySegmentWrite(data_seg, (const char *)&hashResult, sizeof(HashDigest));
-            }
-        }
-        BufferDestroy(&all_dependent_nodes, heap);
-    }
-
-    bool CompileDagDerived(const Frozen::Dag* dag_, MemAllocHeap* heap_, MemAllocLinear* scratch_, const char* dagderived_filename)
-    {
-        heap = heap_;
-        scratch = scratch_;
-        dag = dag_;
-        writer = &_writer;
-        BinaryWriterInit(writer, heap);
-
-        HashTableInit(&shared_strings, heap);
-
-        main_seg = BinaryWriterAddSegment(writer);
-        data_seg = BinaryWriterAddSegment(writer);
-        arraydata_seg = BinaryWriterAddSegment(writer);
-        arraydata2_seg = BinaryWriterAddSegment(writer);
-        leafnodearray_seg = BinaryWriterAddSegment(writer);
-        scannerindex_seg = BinaryWriterAddSegment(writer);
-        str_seg = BinaryWriterAddSegment(writer);
-
-        BinarySegmentWriteUint32(main_seg, Frozen::DagDerived::MagicNumber);
-        node_count = dag->m_NodeCount;
-        BinarySegmentWriteUint32(main_seg, node_count);
-
-        WriteFrozenArrayOfBackLinks();
-
-        WriteLeafInputsAndScannerIndicesToFilesToScan();
-
-        WriteLeafInputHashOffline();
-        BinarySegmentWriteUint32(main_seg, Frozen::DagDerived::MagicNumber);
-
-        HashTableDestroy(&shared_strings);
-
-        bool result = BinaryWriterFlush(writer, dagderived_filename);
-        BinaryWriterDestroy(writer);
-        return result;
-    }
-};
-
-bool CompileDagDerived(const Frozen::Dag* dag, MemAllocHeap* heap, MemAllocLinear* scratch, const char* dagderived_filename)
-{
-    CompileDagDerivedWorker worker;
-    return worker.CompileDagDerived(dag,heap,scratch,dagderived_filename);
-};
-
 static bool CompileDag(const JsonObjectValue *root, BinaryWriter *writer, MemAllocHeap *heap, MemAllocLinear *scratch)
 {
     HashTable<CommonStringRecord, kFlagCaseSensitive> shared_strings;
